Persist Data
Use NSCoding as the data persistence mechanism. Implement the behavior in the Place class to save and load the place. Using the NSCoding approach, the Place class is in charge of storing and loading each of its properties. It needs to save its data by assigning the value of each property to a particular key, and load the data by looking up the information associated with that key.

Create PlaceCacheManager protocol to archive and unarchive place object from the disk.